Add compact control stack and multi-panel split view to embeddings tab
Redesigns the Map/Color by/Tooltip fields toolbar into a compact
in-canvas control stack, and adds a Split view feature (up to 4
independent panels, each with its own map/color-by/tooltip selection,
fully synced to the URL via fixed per-panel param slots).

EmbeddingsTab.tsx becomes a thin wrapper managing panel count;
EmbeddingsPanel.tsx (renamed from the old EmbeddingsTab) holds the
per-panel state, parameterized by panelIndex for URL param naming.
EmbeddingDeckGLVisualization gets a renderControls render-prop so the
new EmbeddingControlStack component can replace the old toolbar
buttons without touching its mouse/lasso/export internals.
