+++ src/pages/studyView/StudyViewURLWrapper.ts
@@ -18,6 +18,17 @@ export default class StudyViewURLWrapper extends URLWrapper<
         | 'plots_vert_selection'
         | 'plots_coloring_selection'
         | 'embeddings_coloring_selection'
+        | 'embeddings_map'
+        | 'embeddings_tooltip_fields'
+        | 'embeddings_panel2_coloring_selection'
+        | 'embeddings_panel2_map'
+        | 'embeddings_panel2_tooltip_fields'
+        | 'embeddings_panel3_coloring_selection'
+        | 'embeddings_panel3_map'
+        | 'embeddings_panel3_tooltip_fields'
+        | 'embeddings_panel4_coloring_selection'
+        | 'embeddings_panel4_map'
+        | 'embeddings_panel4_tooltip_fields'
         | 'geneset_list'
         | 'generic_assay_groups'
     >
@@ -31,6 +42,26 @@ export default class StudyViewURLWrapper extends URLWrapper<
                 isSessionProp: false,
                 nestedObjectProps: PlotsColoringParamProps,
             },
+            embeddings_map: { isSessionProp: false },
+            embeddings_tooltip_fields: { isSessionProp: false },
+            embeddings_panel2_coloring_selection: {
+                isSessionProp: false,
+                nestedObjectProps: PlotsColoringParamProps,
+            },
+            embeddings_panel2_map: { isSessionProp: false },
+            embeddings_panel2_tooltip_fields: { isSessionProp: false },
+            embeddings_panel3_coloring_selection: {
+                isSessionProp: false,
+                nestedObjectProps: PlotsColoringParamProps,
+            },
+            embeddings_panel3_map: { isSessionProp: false },
+            embeddings_panel3_tooltip_fields: { isSessionProp: false },
+            embeddings_panel4_coloring_selection: {
+                isSessionProp: false,
+                nestedObjectProps: PlotsColoringParamProps,
+            },
+            embeddings_panel4_map: { isSessionProp: false },
+            embeddings_panel4_tooltip_fields: { isSessionProp: false },
             geneset_list: { isSessionProp: true },
             generic_assay_groups: { isSessionProp: false },
         });
